refactored  BasicSteps
